ngfw-14979 setupselect added to manage compoents
